+++ data/crac-io/crac-io-json/src/main/java/com/farao_community/farao/data/crac_io_json/deserializers/ContingencyDeserializer.java
@@ -1,0 +1,84 @@
+/*
+ * Copyright (c) 2020, RTE (http://www.rte-france.com)
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+
+package com.farao_community.farao.data.crac_io_json.deserializers;
+
+import com.farao_community.farao.commons.FaraoException;
+import com.farao_community.farao.data.crac_api.NetworkElement;
+import com.farao_community.farao.data.crac_impl.ComplexContingency;
+import com.farao_community.farao.data.crac_impl.SimpleCrac;
+import com.fasterxml.jackson.core.JsonParser;
+import com.fasterxml.jackson.core.JsonToken;
+import com.fasterxml.jackson.core.type.TypeReference;
+
+import java.io.IOException;
+import java.util.ArrayList;
+import java.util.HashSet;
+import java.util.Set;
+
+import static com.farao_community.farao.data.crac_io_json.deserializers.DeserializerNames.*;
+
+/**
+ * @author Baptiste Seguinot {@literal <baptiste.seguinot at rte-france.com>}
+ */
+class ContingencyDeserializer {
+
+    private ContingencyDeserializer() {}
+
+    static void deserialize(JsonParser jsonParser, SimpleCrac simpleCrac) throws IOException {
+        // cannot be done in a standard ComplexContingency deserializer as it requires the simpleCrac to
+        // compare the NetworkElement ids of the ComplexContingency with the NetworkElement of the SimpleCrac
+
+        while(jsonParser.nextToken() != JsonToken.END_ARRAY) {
+
+            String id = null;
+            String name = null;
+            ArrayList<String> networkElementsIds = new ArrayList<>();
+
+            while (!jsonParser.nextToken().isStructEnd()) {
+                switch (jsonParser.getCurrentName()) {
+
+                    case TYPE:
+                        if (!jsonParser.nextTextValue().equals(COMPLEX_CONTINGENCY_TYPE)) {
+                            throw new FaraoException(String.format("SimpleCrac cannot deserialize other contingencies types than %s", COMPLEX_CONTINGENCY_TYPE));
+                        }
+                        break;
+
+                    case ID:
+                        id = jsonParser.nextTextValue();
+                        break;
+
+                    case NAME:
+                        name = jsonParser.nextTextValue();
+                        break;
+
+                    case NETWORK_ELEMENTS:
+                        jsonParser.nextToken();
+                        networkElementsIds = jsonParser.readValueAs(new TypeReference<ArrayList<String>>() {
+                        });
+                        break;
+
+                    default:
+                        throw new FaraoException("Unexpected field: " + jsonParser.getCurrentName());
+                }
+            }
+
+            //add contingency in Crac
+            Set<NetworkElement> networkElements = new HashSet<>();
+            networkElementsIds.forEach(neId -> {
+                NetworkElement ne = simpleCrac.getNetworkElement(neId);
+                if (ne == null) {
+                    throw new FaraoException(String.format("The network element [%s] mentioned in the contingencies is not defined", neId));
+                }
+                networkElements.add(ne);
+            });
+
+            simpleCrac.addContingency(new ComplexContingency(id, name, networkElements));
+
+        }
+    }
+}
