+++ data/crac-io/crac-io-json/src/main/java/com/farao_community/farao/data/crac_io_json/deserializers/DeserializerNames.java
@@ -1,0 +1,49 @@
+/*
+ * Copyright (c) 2020, RTE (http://www.rte-france.com)
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+
+package com.farao_community.farao.data.crac_io_json.deserializers;
+
+/**
+ * @author Baptiste Seguinot {@literal <baptiste.seguinot at rte-france.com>}
+ */
+class DeserializerNames {
+    static final String NETWORK_ELEMENT = "networkElement";
+    static final String NETWORK_ELEMENTS = "networkElements";
+    static final String ID = "id";
+    static final String NAME = "name";
+    static final String OPERATOR = "operator";
+    static final String CONTINGENCY = "contingency";
+    static final String CONTINGENCIES = "contingencies";
+    static final String COMPLEX_CONTINGENCY_TYPE = "complex-contingency";
+    static final String STATE = "state";
+    static final String STATES = "states";
+    static final String SIMPLE_STATE_TYPE = "simple-state";
+    static final String INSTANT = "instant";
+    static final String INSTANTS = "instants";
+    static final String USAGE_METHOD = "usageMethod";
+    static final String SETPOINT = "setpoint";
+    static final String TYPE = "type";
+    static final String CNECS = "cnecs";
+    static final String SIMPLE_CNEC_TYPE = "simple-cnec";
+    static final String THRESHOLD = "threshold";
+    static final String EXTENSIONS = "extensions";
+    static final String RANGE_ACTIONS = "rangeActions";
+    static final String PST_WITH_RANGE_TYPE = "pst-with-range";
+    static final String RANGES = "ranges";
+    static final String USAGE_RULES = "usageRules";
+    static final String FREE_TO_USE_TYPE = "free-to-use";
+    static final String CNEC = "cnec";
+    static final String ON_CONSTRAINT_TYPE = "on-constraint";
+    static final String ON_CONTINGENCY_TYPE = "on-contingency";
+    static final String TOPOLOGY_TYPE = "topology";
+    static final String PST_SETPOINT_TYPE = "pst-setpoint";
+    static final String ACTION_TYPE = "actionType";
+    static final String COMPLEX_NETWORK_ACTION_TYPE = "complex-network-action";
+    static final String ELEMENTARY_NETWOK_ACTIONS = "elementaryNetworkActions";
+    static final String NETWORK_ACTIONS = "networkActions";
+    static final String ALIGNED_RANGE_ACTIONS_TYPE = "aligned-range-action";
+}
