+++ data/crac-io/crac-io-json/src/test/java/com/farao_community/farao/data/crac_io_json/RoundTripUtil.java
@@ -7,6 +7,7 @@
 package com.farao_community.farao.data.crac_io_json;
 
 import com.farao_community.farao.data.crac_impl.SimpleCrac;
+import com.farao_community.farao.data.crac_io_json.deserializers.SimpleCracDeserializer;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.ObjectWriter;
